elt-hub-core/pyproject.toml
[project]
name = "elt-hub-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "polars>=1.35.2",
    "psycopg[binary,pool]>=3.2.13",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
